Fix progress reporting in qvm-convert-img-gnome
qimg-convert-client needs to write to the original stdout (of
qrexe-client-vm process), not its own. Its own stdout is connected to
the DispVM and is used for the actual image data.

It looks like it have never worked before...

Also, report 10% at the script startup, as it means the DispVM has been
started (which takes the most time).